[Add] Favorite Anime - Added WithAnimeSearch trait - Updated WithSearch trait - Added favorite anime page - Added favorite anime section to profile - Moved profile/anime-library to profile/library/anime/index
