+++ resources/views/livewire/profile/library/anime/index.blade.php
